+++ docs/fr_FR/Changelog.rst
@@ -2,6 +2,7 @@ ChangeLog
 =========
 
 - Ikea capteur qualité de l'air VINDSTYRKA: Mise-à-jour du modèle (2681).
+- Interne: Cmd: Amélioration formatAttribute() pour type x39/single precision + int8/int24/int32.
 
 240116-BETA-1
 -------------